requirements.txt
Adafruit-Blinka==8.66.2
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-neopixel==6.3.17
adafruit-circuitpython-pixelbuf==2.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.84.1
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
sysv_ipc==1.1.0
typing_extensions==4.15.0
